requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
aiohttp==3.8.3
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
charset-normalizer==2.1.1
    # via aiohttp
discord-py==2.1.0
    # via mashmaestro (pyproject.toml)
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.24.1
    # via pandas
pandas==1.5.3
    # via mashmaestro (pyproject.toml)
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.21.1
    # via mashmaestro (pyproject.toml)
pytz==2022.7.1
    # via pandas
six==1.16.0
    # via python-dateutil
tabulate==0.9.0
    # via mashmaestro (pyproject.toml)
wheel==0.38.4
    # via mashmaestro (pyproject.toml)
yarl==1.8.2
    # via aiohttp
